requirements.txt
streamlit==1.35.0
requests==2.32.3
Pillow==10.3.0
streamlit-lottie==0.0.5
